+++ R/geoflow_handler_contact.R
@@ -71,10 +71,7 @@ list_contact_handlers <- function(raw = FALSE){
 #'
 list_contact_handler_options <- function(id, raw = FALSE){
   out <- NULL
-  handlers <- list_contact_handlers(raw = TRUE)
-  handler <- handlers[sapply(handlers, function(x){x$id == id})]
-  if(length(handler)==0) stop(sprintf("No handler with id '%s'!", id))
-  handler <- handler[[1]]
+  handler <- get_contact_handler(id)
   if(raw) return(handler$available_options)
   if(length(handler$available_options)>0){
     out <- data.frame(
@@ -88,4 +85,51 @@ list_contact_handler_options <- function(id, raw = FALSE){
     out <- data.frame(name = character(0), definition = character(0))
   }
   return(out)
+}
+
+#' @name get_contact_handler
+#' @aliases get_contact_handler
+#' @title get_contact_handler
+#' @description \code{get_contact_handler} allows to get a contact handler
+#' 
+#' @usage get_contact_handler(id)
+#' 
+#' @param id A contact handler identifier
+#' @return an object of class \link{geoflow_handler}
+#' @author Emmanuel Blondel, \email{emmanuel.blondel1@@gmail.com}
+#' @export
+#'
+get_contact_handler <- function(id){
+  handlers <- list_contact_handlers(raw = TRUE)
+  handler <- handlers[sapply(handlers, function(x){x$id == id})]
+  if(length(handler)==0) stop(sprintf("No handler with id '%s'!", id))
+  handler <- handler[[1]]
+  return(handler)
+}
+
+#' @name read_contacts
+#' @aliases read_contacts
+#' @title read_contacts
+#' @description \code{read_contacts} allows to read contacts
+#' 
+#' @usage read_contacts(id, source, config)
+#' 
+#' @param id a contact handler identifier
+#' @param source source
+#' @param config a geoflow config (output of \link{initWorkflow}). Default is \code{NULL}
+#' @return a list of object of class \link{geoflow_contact}
+#' @author Emmanuel Blondel, \email{emmanuel.blondel1@@gmail.com}
+#' @export
+#' 
+read_contacts <- function(id, source, config = NULL){
+  handler <- get_contact_handler(id)
+  if(is.null(config)){
+    config = add_config_logger(list())
+  }
+  handler$fun(
+    handler = handler,
+    source = source,
+    config = config,
+    handle = TRUE
+  )
 }